requirements.txt
streamlit
Pillow
python-dotenv
google-generativeai
SpeechRecognition
pyaudio
